feat: Se agrego una comprobacion que tira un mensaje si no se encuentra un empleado al momento de buscar

+++ AplicacionRepuestos/src/main/java/com/gabrielcalderon/RepuestosAutomotrices/service/EmpleadoServiceImplements.java
@@ -19,14 +19,21 @@ public class EmpleadoServiceImplements implements EmpleadoService{
     public List<Empleado> getAllEmpleados() {
         return empleadoRepository.findAll();
     }
+
     @Override
     public Empleado getEmpleadoById(Integer id) {
+        Empleado empleado = empleadoRepository.findById(id).orElse(null);
+        if (empleado == null) {
+            throw new IllegalArgumentException("Empleado no encontrado");
+        }
         return empleadoRepository.findById(id).orElse(null);
     }
+
     @Override
     public Empleado saveEmpleado(Empleado empleado) throws RuntimeException {
         return empleadoRepository.save(empleado);
     }
+
     @Override
     public Empleado updateEmpleado(Integer id, Empleado empleado) {
         Empleado empleado1 = empleadoRepository.findById(id).orElse(null);
@@ -35,6 +42,7 @@ public class EmpleadoServiceImplements implements EmpleadoService{
             empleado1.setApellido_empleado(empleado.getApellido_empleado());
             empleado1.setPuesto_empleado(empleado.getPuesto_empleado());
             empleado1.setEmail_empleado(empleado.getEmail_empleado());
+
         } else {
             throw new IllegalArgumentException("Empleado no encontrado");
         }
@@ -42,7 +50,11 @@ public class EmpleadoServiceImplements implements EmpleadoService{
     }
 
     @Override
-    public void deleateEmpleado(Integer id) {
+    public void deleteEmpleado(Integer id) {
+        Empleado empleado = empleadoRepository.findById(id).orElse(null);
+        if (empleado == null) {
+            throw new IllegalArgumentException("Empleado no encontrado");
+        }
         empleadoRepository.deleteById(id);
     }
 }
